Finalizando a classe Estacionamento

+++ Entrega 2/src/trabalhoOO/Main.java
@@ -84,7 +84,7 @@ public class Main {
                     float mensalidade = Float.parseFloat(JOptionPane.showInputDialog(null, "Insira o valor da mensalidade:", "Cadastrar Estacionamento", JOptionPane.PLAIN_MESSAGE));
                     float valorEvento = Float.parseFloat(JOptionPane.showInputDialog(null, "Insirao valor do evento:", "Cadastrar Estacionamento", JOptionPane.PLAIN_MESSAGE));
                     
-                    float valorContratantePct = valorContratante / 100; 
+                    float valorContratantePct = (float) valorContratante / 100;
 
 
                     Estacionamento estacionamentoNovo = new Estacionamento(nomeEstacionamento, capacidade, horarioAbertura, horarioEncerramento, valorContratantePct);
